Customizer: When a block theme is active, add an information about Site Editor in the Customizer.
This change adds an information notice to the customizer when a block theme is active and the customizer is also available (for example when a plugin activates it), to encourage people to use the Site Editor for the best full site customization experience.

Props ironprogrammer, antonvlasenko, Clorith, audrasjb, psmits1567, tobifjellner, costdev, webcommsat, joedolson, pbiron.
Merges [53024] and [53025] to the 5.9 branch.
Fixes #54939.

Built from https://develop.svn.wordpress.org/branches/5.9@53026


git-svn-id: https://core.svn.wordpress.org/branches/5.9@52615 1a063a9b-81f0-0310-95a4-ce76da25c4cd