Help/About: Add cache busting string to video thumbnail.
Append `?ver=6.0` to the image used for linking to the video to break the cache in browsers, proxy servers and on the CDN.

A new version of the image has been uploaded and replaced the file in the original location.

Props annezazu, critterverse, ryelle, costdev, ironprogrammer, dd32, peterwilsoncc.
Merges [53447] to the 6.0 branch.
Fixes #55808.
Built from https://develop.svn.wordpress.org/branches/6.0@53649


git-svn-id: https://core.svn.wordpress.org/branches/6.0@53208 1a063a9b-81f0-0310-95a4-ce76da25c4cd