Themes: Revert the change of theme preview width on Add Themes screen.
The intention was for the preview ratio to remain accurate for the user's screen size when the sidebar is open, but this appears to have caused unintended side effects to the preview alignment.

Follow-up to [52647].

Props superbthemescom, Presskopp.
Merges [53663] to the 6.0 branch.
Fixes #56097. See #54764.
Built from https://develop.svn.wordpress.org/branches/6.0@53664


git-svn-id: https://core.svn.wordpress.org/branches/6.0@53223 1a063a9b-81f0-0310-95a4-ce76da25c4cd